adds clown glue

+++ code/modules/paperwork/photocopier.dm
@@ -268,29 +268,29 @@
 		copying = 0
 	if(istype(O, /obj/item/weapon/paper))
 		if(copier_empty())
-			user.drop_item(O, src)
-			copy = O
-			user << "<span class='notice'>You insert [O] into [src].</span>"
-			flick("bigscanner1", src)
-			updateUsrDialog()
+			if(user.drop_item(O, src))
+				copy = O
+				user << "<span class='notice'>You insert [O] into [src].</span>"
+				flick("bigscanner1", src)
+				updateUsrDialog()
 		else
 			user << "<span class='notice'>There is already something in [src].</span>"
 	else if(istype(O, /obj/item/weapon/photo))
 		if(copier_empty())
-			user.drop_item(O, src)
-			photocopy = O
-			user << "<span class='notice'>You insert [O] into [src].</span>"
-			flick("bigscanner1", src)
-			updateUsrDialog()
+			if(user.drop_item(O, src))
+				photocopy = O
+				user << "<span class='notice'>You insert [O] into [src].</span>"
+				flick("bigscanner1", src)
+				updateUsrDialog()
 		else
 			user << "<span class='notice'>There is already something in [src].</span>"
 	else if(istype(O, /obj/item/device/toner))
 		if(toner <= 0)
-			user.drop_item(O)
-			qdel(O)
-			toner = 40
-			user << "<span class='notice'>You insert [O] into [src].</span>"
-			updateUsrDialog()
+			if(user.drop_item(O))
+				qdel(O)
+				toner = 40
+				user << "<span class='notice'>You insert [O] into [src].</span>"
+				updateUsrDialog()
 		else
 			user << "<span class='notice'>This cartridge is not yet ready for replacement! Use up the rest of the toner.</span>"
 	else if(istype(O, /obj/item/weapon/wrench))
